+++ DashAI/front/src/styles/theme.js
@@ -1,4 +1,6 @@
 const theme = {
+
+
   rootBackground: "#2e3037",
   title: "#fff",
   subTitle: "#fff",