+++ DashAI/front/src/tabs/Play.jsx
@@ -1,13 +1,8 @@
 import React from "react";
-import { CustomContainer } from "../styles/globalComponents";
 import PlayCard from "../components/PlayCard";
 
 function Play() {
-  return (
-    <CustomContainer>
-      <PlayCard />
-    </CustomContainer>
-  );
+  return <PlayCard />;
 }
 
 export default Play;